Añadido verificación tipo de fichero (formato de video solamente)

+++ Classify_Series.py
@@ -3,17 +3,19 @@ import os
 import time
 import sys
 
-DIR = "/Vault/series/"
+DIR = "/Volumes/Vault/Series/"
 LOGFILE = "/Volumes/Vault/flexget/flexget.log"
 
 nombre_serie = os.listdir(DIR)
 fobj = sys.stdout
+ftype_reg_exp = re.compile('.*\.(avi|mkv|mp4)$')
 
 def escribe_log(tipo, mensaje):
     log_msg = time.strftime("[%d/%m/%Y %H:%M:%S] CLASSIFY.")
     log_msg += tipo + " "
     log_msg += mensaje + "\n"
     fobj.write(log_msg)
+
 
 try:
     fobj = open(LOGFILE, 'a')
@@ -26,19 +28,21 @@ for serie in nombre_serie:
     if os.path.isdir(cwd):
         episodios = os.listdir(cwd)
         for item in episodios:
-            if os.path.isfile(cwd + item):
+            escribe_log("INFO", "Procesando directorio " + cwd)
+            if os.path.isfile(cwd + item) and (ftype_reg_exp.match(cwd + item)):
                 temporada = re.findall("S|s[0-9]{2}", item)[0]
                 try:
-                    if os.access(cwd,os.W_OK):
+                    if os.access(cwd, os.W_OK):
                         os.rename(cwd + item, cwd + temporada.upper() + '/' + item)
-                        escribe_log("INFO","Episodio " + item + " movido a " + temporada.upper())
+                        escribe_log("INFO", "Episodio " + item + " movido a " + temporada.upper())
                     else:
-                        escribe_log("ERROR","Error de permisos, imposible mover fichero")
+                        escribe_log("ERROR", "Error de permisos, imposible mover fichero")
                 except OSError, e:
                     if e.errno == 2:
                         if os.access(cwd, os.W_OK):
                             os.mkdir(cwd + temporada.upper())
                             os.rename(cwd + item, cwd + temporada.upper() + '/' + item)
-                            escribe_log("INFO","Episodio " + item + " movido a " + temporada.upper() + ", directorio " + temporada.upper() + " creado")
+                            escribe_log("INFO",
+                                        "Episodio " + item + " movido a " + temporada.upper() + ", directorio " + temporada.upper() + " creado")
                         else:
-                            escribe_log("ERROR","Error de permisos, imposible crear directorio")
+                            escribe_log("ERROR", "Error de permisos, imposible crear directorio")
